Zip archive doesn't work if I package it on Windows.
The root cause is zip archive from Mac strat with directory.
However, when I archive on the Windows, it could start with file.
The file has intermediate directory, if is not created, the open file failes.
That is why I create the intermediate directory. It works on Mac. I also try Windows as well.
I commit this for share with windows.